Fix check in knots function
Previously checked that input size was equal to knot vector size first
(L93) and then that input size was equal to number of break points.
Since number of knots != number of break points, this is impossible to
satisfy.  Changed to only check that input size is equal to number of
break points, which is consistent with behavior of underlying GSL
function gsl_bspline_knots, which expects a vector of break point
locations to be passed.